backend/requirements.txt
fastapi==0.60.1
pillow==7.1.2
tensorflow==2.4.1
python-multipart==0.0.5 
aiofiles==0.4.0
matplotlib==3.3.4
numpy==1.19.5